Update config.yml
Fixed the identifiers for the mad scientist lab for both creation and recipes. The MONSTER_EGG idenitifer refers to the strone brick blocks that spawn sliverfish when broken, whereas the SPAWN_EGG refers to spawn eggs. The durability values seem to be correct so I did not mess with those.

http://minecraft.gamepedia.com/Monster_Egg

http://minecraft.gamepedia.com/Spawn_Egg